+++ packages/odata-service-inquirer/src/index.ts
@@ -43,7 +43,7 @@ import { getPromptHostEnvironment, PromptState } from './utils';
  * @param enableGuidedAnswers - if true, the prompts will use guided answers to help users with validation errors
  * @param telemetryClient - the telemetry client to use for sending telemetry data
  * @param isYUI - if true, the prompt is being called from the Yeoman UI extension host
- * @param connectedSystem - if available passing an already connected system connection will prevent re-authentication for re-entrance ticket and service keys connection types
+ * @param connectedSystem - if available passing an already connected system connection will prevent re-authentication for re-entrance ticket connection types
  * @returns the prompts used to provide input for odata service generation and a reference to the answers object which will be populated with the user's responses once `inquirer.prompt` returns
  */
 async function getPrompts(
@@ -133,7 +133,7 @@ function getEntityRelatedPrompts(
  * @param enableGuidedAnswers - if true, the prompts will use guided answers to help users with validation errors
  * @param telemetryClient - the telemetry client to use for sending telemetry data
  * @param isYUI - if true, the prompt is being called from the Yeoman UI extension host
- * @param connectedSystem - if available passing an already connected system connection will prevent re-authentication for re-entrance ticket and service keys connection types
+ * @param connectedSystem - if available passing an already connected system connection will prevent re-authentication for re-entrance ticket connection types
  * @returns the prompt answers
  */
 async function prompt(